Move PDF interfaces
Move PDF interfaces inside existing source code file instead of creating new file just for them
